remove warning on ubuntu

+++ tools/hitag2crack/common/ht2crackutils.h
@@ -16,7 +16,7 @@
 
 #define HEX_PER_ROW 16
 
-void writebuf(unsigned char *buf, uint64_t val, unsigned int len);
+void writebuf(unsigned char *buf, uint64_t val, uint16_t len);
 void shexdump(unsigned char *data, int data_len);
 void printbin(unsigned char *c);
 void printbin2(uint64_t val, unsigned int size);
